Revert "Render rework"
