+++ frontend/components/Leaderboard_item.jsx
@@ -9,7 +9,7 @@ const Leaderboard_item = ({ id, rank, expanded, setExpanded }) => {
   const address = '0xbc4ca0eda7647a8ab7c2061c2e118a18a936f13d';
 
   const getCityData = async () => {
-    let element = await (await fetch(`http://0.0.0.0:8000/cities/${id}/data`)).json();
+    let element = await (await fetch(`http://localhost:8000/cities/${id}/data`)).json();
     // console.log(id, element);
     setCity(element);
   }
