--- frontend/components/game/GridData.js
@@ -1,5 +1,0 @@
-const gridDimensions = 100;
-const gridSize = 10;
-const plotSize = gridDimensions/gridSize;
-
-export {gridDimensions,gridSize,plotSize}
