+++ frontend/next.config.js
@@ -4,7 +4,4 @@ const nextConfig = {
   swcMinify: false
 }
 
-const withTM = require('next-transpile-modules')(['three'])
-module.exports = withTM()
-
 module.exports = nextConfig
