+++ frontend/package.json
@@ -9,22 +9,17 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@react-spring/three": "^9.4.4",
-    "@react-three/drei": "^8.20.2",
-    "@react-three/fiber": "^7.0.26",
     "dat.gui": "^0.7.9",
     "ethers": "^5.5.4",
     "next": "12.1.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "three": "^0.139.2",
-    "zustand": "^3.7.1",
     "react-scrollbar-size": "^4.0.0",
     "spinners-react": "^1.0.6",
+    "three": "^0.138.3"
   },
   "devDependencies": {
     "eslint": "8.9.0",
-    "eslint-config-next": "12.1.0",
-    "next-transpile-modules": "^9.0.0"
+    "eslint-config-next": "12.1.0"
   }
 }
