+++ game_render/grid.js
@@ -1,0 +1,16 @@
+class gridObject{
+    constructor(buildingType,coordinate,coordinate1){
+        this.buildingType=buildingType;
+        this.coordinate=coordinate;
+        this.coordinate1=coordinate1;
+    }
+
+}
+
+class grid {
+    constructor(height,width){
+        this.height=height;
+        this.width=width;
+    }
+
+}
