+++ game_render/main.js
@@ -291,6 +291,27 @@ document.addEventListener(
     // console.log(buildingModule.buildingList);
     // console.log("built: ",type);
   }
+
+  /// FUNCTION  TO REMOVE A BUILDING IN THE MAP
+  function demolish(x,y)
+  {
+    let occupied=find(new buildingModule.Coordinate(x,y),new buildingModule.Coordinate(x,y))
+    if(!(occupied===undefined)) //IF OCCUPIED
+    {
+      scene.children.forEach(element => {
+        if(element.gridx)
+        {
+          if(element.gridx==occupied.start.x && element.gridy==occupied.start.y && element.name!='gridSquare')
+          {
+            scene.remove(element);
+          }
+        }
+      });
+      buildingModule.buildingList.splice(buildingModule.buildingList.findIndex(element => element==occupied),1);
+    }
+
+  }
+
   function sceneAdd(start,end,elementType,level){
     loader.load( './models/' + elementType + '_placeholder.glb', function ( gltf ) {
       const model = gltf.scene.children[0];
@@ -313,26 +334,6 @@ document.addEventListener(
     }, undefined, function ( error ) {
       console.error( error );
     } );
-}
-
-/// FUNCTION  TO REMOVE A BUILDING IN THE MAP
-function demolish(x,y)
-{
-  let occupied=find(new buildingModule.Coordinate(x,y),new buildingModule.Coordinate(x,y))
-  if(!(occupied===undefined)) //IF OCCUPIED
-  {
-    scene.children.forEach(element => {
-      if(element.gridx)
-      {
-        if(element.gridx==occupied.start.x && element.gridy==occupied.start.y && element.name!='gridSquare')
-        {
-          scene.remove(element);
-        }
-      }
-    });
-    buildingModule.buildingList.splice(buildingModule.buildingList.findIndex(element => element==occupied),1);
-  }
-
 }
 
 /// FUNCTION TO FIND A BUILDING, IF IT EXISTS --- RETURNS UNDEFINED IF DOESNT EXIST
